+++ src/main/java/car_racing/model/Car.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.model;
 
 public class Car implements RacingPlayer {
     private final String name;
