+++ src/main/java/car_racing/model/GamePlayStrategy.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.model;
 
 public interface GamePlayStrategy {
     Integer generateScore();
