+++ src/main/java/car_racing/model/GameRule.java
@@ -1,13 +1,10 @@
-package car_racing;
+package car_racing.model;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 
 public class GameRule {
-
-    private GameRule() {}
-
     private static final Integer STANDARD_SCORE = 4;
 
     private static final GamePlayStrategy gp = new RandomPlayStrategy();
@@ -20,4 +17,6 @@ public class GameRule {
         Integer maxScore = players.stream().max(Comparator.comparingInt(RacingPlayer::getScore)).get().getScore();
         return players.stream().filter(p -> Objects.equals(p.getScore(), maxScore)).toArray(RacingPlayer[]::new);
     }
+
+    private GameRule() {}
 }
