+++ src/main/java/car_racing/model/RacingPlayer.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.model;
 
 public interface RacingPlayer {
     void proceedNextTurn(Boolean proceed);
