+++ src/test/java/car_racing/GamePlayStrategyTest.java
@@ -1,5 +1,7 @@
 package car_racing;
 
+import car_racing.model.GamePlayStrategy;
+import car_racing.model.RandomPlayStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
